# docs/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: 寝室电费学生端 API
  version: 0.7.0
  description: T0 冻结契约。金额为十进制字符串，未知为 null，日期为上海含首尾日期，时间戳带时区。202 仅表示持久受理。业务服务按 x-stage
    实现。
servers:
- url: /api/v1
paths:
  /room-bindings/{id}:
    delete:
      operationId: unbind_room
      tags:
      - room
      summary: unbind_room
      x-owner: room
      x-stage: T3
      security:
      - SessionCookie: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Origin
        in: header
        required: true
        schema:
          type: string
          format: uri
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: 持久受理，按查询接口跟踪
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/AcceptedOperation'
                  meta:
                    $ref: '#/components/schemas/Meta'
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
      x-result-query: /operations/{id}
      x-persistence: room.room_operations
    get:
      operationId: get_binding
      tags:
      - room
      summary: get_binding
      x-owner: room
      x-stage: T3
      security:
      - SessionCookie: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Binding'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
...
  /auth/school-credential:
    delete:
      operationId: revoke_credential
      tags:
      - identity
      summary: revoke_credential
      x-owner: identity
      x-stage: T3
      security:
      - SessionCookie: []
      parameters:
      - name: Origin
        in: header
        required: true
        schema:
          type: string
          format: uri
      - $ref: '#/components/parameters/CSRF'
      responses:
        '202':
          description: 持久受理，按查询接口跟踪
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/AcceptedOperation'
                  meta:
                    $ref: '#/components/schemas/Meta'
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      x-version-required: expected_version；缺失 428，冲突 409
      x-result-query: /operations/{id}
      x-persistence: identity.credential_operations
...
  /room-bindings:
    get:
      operationId: list_bindings
      tags:
      - room
      summary: list_bindings
      x-owner: room
      x-stage: T2
      security:
      - SessionCookie: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 128
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Bindings'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
    post:
      operationId: bind_room
      tags:
      - room
      summary: bind_room
      x-owner: room
      x-stage: T3
      security:
      - SessionCookie: []
      parameters:
      - name: Origin
        in: header
        required: true
        schema:
          type: string
          format: uri
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/AcceptedOperation'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '202':
          description: 持久受理，按查询接口跟踪
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/AcceptedOperation'
                  meta:
                    $ref: '#/components/schemas/Meta'
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
        '400':
          description: CAPTCHA_INVALID, CAPTCHA_EXPIRED, SNAPSHOT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: APP_SESSION_EXPIRED, SCHOOL_LOGIN_REJECTED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF_REJECTED, ORIGIN_REJECTED, PAYMENT_UNAVAILABLE, FEATURE_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, OPERATION_IN_PROGRESS,
            SCHOOL_REAUTH_REQUIRED, REAUTH_ACCOUNT_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410':
          description: ROOM_CANDIDATE_EXPIRED, SNAPSHOT_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: INVALID_ARGUMENT, INVALID_INTERVAL, INVALID_DATE_RANGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: PRECONDITION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: SCHOOL_PROTOCOL_CHANGED, SCHOOL_INVALID_RESPONSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: SCHOOL_UNAVAILABLE, CIRCUIT_OPEN, DEPENDENCY_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: SCHOOL_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-cache-control: no-store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindRequest'
      x-result-query: /operations/{id}
      x-persistence: room.room_operations
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-elect_session
      description: Secure; HttpOnly; SameSite=Lax; Path=/；不设置 Domain
  parameters:
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 按用户+操作类型分区，原请求摘要一致才能重放，至少保留 180 天
      schema:
        type: string
        minLength: 16
        maxLength: 128
  schemas:
...
    Meta:
      additionalProperties: false
      properties:
        request_id:
          format: uuid
          title: Request Id
          type: string
        server_time:
          format: date-time
          title: Server Time
          type: string
      required:
      - request_id
      - server_time
      title: Meta
      type: object
...
    Error:
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          title: Message
          type: string
        retryable:
          title: Retryable
          type: boolean
        retry_after_seconds:
          anyOf:
          - minimum: 0
            type: integer
          - type: 'null'
          title: Retry After Seconds
        requires_reauth:
          title: Requires Reauth
          type: boolean
        field_errors:
          additionalProperties:
            type: string
          title: Field Errors
          type: object
        current_version:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          default: null
          title: Current Version
        existing_operation_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          default: null
          title: Existing Operation Id
      required:
      - code
      - message
      - retryable
      - retry_after_seconds
      - requires_reauth
      - field_errors
      title: Error
      type: object
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/Error'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - error
      - meta
      title: ErrorEnvelope
      type: object
    VersionRequest:
      additionalProperties: false
      properties:
        expected_version:
          minimum: 1
          title: Expected Version
          type: integer
      required:
      - expected_version
      title: VersionRequest
      type: object
...
    OperationSummary:
      additionalProperties: false
      properties:
        id:
          format: uuid
          title: Id
          type: string
        type:
          enum:
          - credential_revoke
          - binding_sync
          - bind_room
          - unbind_room
          - switch_default
          - balance_refresh
          - history_sync
          - qr_refresh
          title: Type
          type: string
        state:
          enum:
          - accepted
          - running
          - reconciling
          - unknown
          - succeeded
          - failed
          - cancelled
          title: State
          type: string
        target_binding_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Target Binding Id
        created_at:
          format: date-time
          title: Created At
          type: string
      required:
      - id
      - type
      - state
      - target_binding_id
      - created_at
      title: OperationSummary
      type: object
...
    AcceptedOperation:
      additionalProperties: false
      properties:
        operation_id:
          format: uuid
          title: Operation Id
          type: string
        state:
          enum:
          - accepted
          - running
          - reconciling
          - unknown
          - succeeded
          - failed
          - cancelled
          title: State
          type: string
        poll_url:
          title: Poll Url
          type: string
      required:
      - operation_id
      - state
      - poll_url
      title: AcceptedOperation
      type: object
...
    Balance:
      additionalProperties: false
      properties:
        amount:
          anyOf:
          - pattern: ^-?(0|[1-9]\d{0,11})\.\d{2}$
            type: string
          - type: 'null'
          title: Amount
        currency:
          const: CNY
          title: Currency
          type: string
        source:
          anyOf:
          - enum:
            - school_bound_rooms
            - school_room_candidate
            type: string
          - type: 'null'
          title: Source
        fetched_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Fetched At
        school_observed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: School Observed At
        stale:
          title: Stale
          type: boolean
        refresh_state:
          enum:
          - idle
          - pending
          - ready
          - failed
          title: Refresh State
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
      required:
      - amount
      - currency
      - source
      - fetched_at
      - school_observed_at
      - stale
      - refresh_state
      - error_code
      title: Balance
      type: object
    Binding:
      additionalProperties: false
      properties:
        id:
          format: uuid
          title: Id
          type: string
        room_id:
          format: uuid
          title: Room Id
          type: string
        building:
          title: Building
          type: string
        number:
          title: Number
          type: string
        display_name:
          title: Display Name
          type: string
        status:
          enum:
          - active
          - rechecking
          - inactive
          title: Status
          type: string
        balance:
          anyOf:
          - $ref: '#/components/schemas/Balance'
          - type: 'null'
      required:
      - id
      - room_id
      - building
      - number
      - display_name
      - status
      - balance
      title: Binding
      type: object
    Bindings:
      additionalProperties: false
      properties:
        page:
          minimum: 1
          title: Page
          type: integer
        page_size:
          maximum: 100
          minimum: 1
          title: Page Size
          type: integer
        total:
          minimum: 0
          title: Total
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Binding'
          title: Items
          type: array
        default_binding_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Default Binding Id
        preference_version:
          minimum: 1
          title: Preference Version
          type: integer
        default_switch_operation_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Default Switch Operation Id
        binding_removal_operation_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          default: null
          title: Binding Removal Operation Id
        preference_state:
          default: ready
          enum:
          - ready
          - switching
          - blocked
          title: Preference State
          type: string
        sync_status:
          enum:
          - ready
          - loading
          - empty
          - stale
          - partial
          - failed
          - unavailable
          title: Sync Status
          type: string
        last_synced_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Synced At
        pending_operations:
          items:
            $ref: '#/components/schemas/OperationSummary'
          maxItems: 20
          title: Pending Operations
          type: array
        pending_operations_truncated:
          title: Pending Operations Truncated
          type: boolean
      required:
      - page
      - page_size
      - total
      - items
      - default_binding_id
      - preference_version
      - default_switch_operation_id
      - sync_status
      - last_synced_at
      - pending_operations
      - pending_operations_truncated
      title: Bindings
      type: object
...
    BindRequest:
      additionalProperties: false
      properties:
        candidate_id:
          maxLength: 128
          minLength: 1
          title: Candidate Id
          type: string
      required:
      - candidate_id
      title: BindRequest
      type: object
...